isl_ast_codegen.c: add_domain: only take single copy of inverse schedule
When this function was introduced in isl-0.15-236-g0ab343912b
(isl_ast_codegen.c: generate_domain: extract out add_domain,
Sun Aug 2 12:25:48 2015 +0200), the caller was performing
a gist on the inverse schedule and passed down both the original version and
the gisted version.
In an earlier commit, the gisting was removed, so these two versions
of the inverse schedule are always the same.
Only pass it in once.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>