detect specific instance of modulos in constraints
During AST generation, when considering a given schedule dimension,
any integer divisions that may depend on this dimension
are removed to be able to extract usable lower and upper bounds.
If the schedule dimension happens to be equal to a modulo
of a floor or modulo expression, then the constraints
for encoding this relationship will involve an integer division
that is defined in terms of the schedule dimension.
Removing this integer division results in an approximation
where the schedule dimension may no longer have a fixed value.
This means that a loop gets generated with a (complicated) condition
in the body that picks out a single iteration of the loop.

Detect a specific instance of this kind of nested modulo expression and
make it explicit in the constraints such that the integer division
no longer depends on the schedule dimension and therefore
does not need to be removed, allowing the AST generator
to detect that the schedule dimension has a fixed value
in terms of the parameters and outer schedule dimensions.

Reviewed-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>