simplify constraints by checking for residues
In particular, if there is a pair of constraints

     f(x) >= 0
    -f(x) + c >= 0

as well as some other constraint

    g(x) >= 0

such that

    g(x) - f(x) = n h(x) + c'

with

    0 <= c' < n - c

then replace this other constraint by

    h(x) >= 0

Integer divisions are simplified in a similar way.
However, constraints defining integer divisions are not modified
since this runs the risk of ending up in an infinite loop
where the constraints are first simplified after which
the integer division definition gets plugged into
them, resulting in the original constraints.

A similar effect can occur in the presence of nested integer divisions.
The situation is more difficult to detect in this case because
it involves multiple substitution steps.
Simply avoid simplifying any constraint with an integer division
involving nested integer divisions and a unit coefficient
after simplification.

This should hopefully avoid all risks of infinite loops.
It would perhaps be even safer to not perform the simplification
if any integer divisions are involved, but then that would
also remove some of the benefits of this simplification.

Also simplifying the integer divisions themselves
is also intended to avoid infinite loops,
but it is too fragile to rely on.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>