isl_map_coalesce: handle empty facet during wrapping
If the tableau representing the facet is empty then
no wrapping can be performed.  When wrapping is required,
this is considered a failure.  Otherwise, it simply
a possibly missed opportunity.

The facet can turn out to be empty in particular
if the bounding constraint contains only rational points.
In the new test case this happens because
equality constraints are only detected during coalescing and
the remaining inequality constraints then do not get
fully simplified resulting in a constraint of the form 2a >= 7,
which gets used as a facet constraint.

Reported-by: Andrei Lascu <alas.20073@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>