isl_coalesce.c: add_selected_wraps_around_facet: add missing rollback
This was missing from the previous commit and would result
in the disjunct with empty facet getting dropped entirely.
To be able to check that no disjunct is being dropped,
test_coalesce_union is extended to check that the coalescing
not only succeeds but that the result is equal to the input.

Reported-by: Andrei Lascu <alas.20073@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>