cpp/cpp-checked.h.top: use consistent indentation for ctx class
In particular, use indentation that is consistent with the stat and size
classes and with the ctx class in cpp/cpp.h.top.
In fact, the two ctx class definitions are identical and will
remain so over the next commits.
It may therefore make sense to only define it once.
However, the surrounding code is different and adding a mechanism
for splicing in the common ctx class may not be worth it.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>