Method::print_cpp_arg_list: extract out Method::on_cpp_arg_list
The correspondence between C function arguments and
tuple kinds is not entirely obvious.
Currently, this correspondence is only needed while printing
the arguments.  However, an upcoming commit will need
to iterate over the C function arguments and the corresponding
tuple kinds before the printing.

Extract out a generic Method::on_cpp_arg_list that can
be used both for printing and for this other iteration
over the arguments.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>