isl_ast_node_dup: fix handling of degenerate for loops
Ever since the introduction of AST generation support
in isl-0.10-199-g5888ac1c9f (add support for generating ASTs
from schedule relations, Mon Sep 17 22:11:06 2012 +0200),
the cond and inc fields have been documented as being NULL
in the case of degenerate for loops.
However this was not being taken into account by isl_ast_node_dup.
In fact, it did not even preserve the flag indicating the property.

Preserve the property and do not copy these fields
in the case of a degenerate for loop.  In particular,
NULL values in the copy should definitely not be considered an error.
Note that all fields are initialized to zero in isl_ast_node_alloc,
so these fields do not need to be explicitly initialized to NULL
once more.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>