add exported isl_ast_node_map_descendant_bottom_up
This complements isl_ast_node_foreach_descendant_top_down,
allowing the user to modify the visited nodes.
In particular, it allows the user to replace
the annotation of some nodes.

The test case is a bit artificial, but there is very little
that can be modified in an AST (other than replacing
the annotation, which is a bit cumbersome to test).

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>