isl_ast_node_dup: preserve annotation
A copy of a node should also have a copy of its annotation.
In practice, this commit does not make any difference
because currently the only way of modifying an AST node
(which could result in a copy being created)
is isl_ast_node_set_annotation, which overwrites
the annotation in the copy anyway.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>