detect some modulo expressions when extracting a function
In the worst case, isl_pw_multi_aff_from_map resorts
to parametric integer programming to find a function representation
of a single valued binary relation, but it detects
some special cases first by examining the constraints.
This prevents a subdivision of the domain that may result
from parametric integer programming.
In particular, it already checks for specific cases
of integer divisions.
Extend this to also check for some forms of modulo expressions.
This ensures that the new test cases can be represented
by functions defined over a single cell.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>