isl_map.c: extract out generic extract_bound_from_constraint
In particular, extract out a generic version of part
of isl_basic_map_inequality_extract_output_upper_bound
that works for both lower and upper bounds and also
equality constraints and that allows for the selection
of an initial subset of the variables.

This generic version will be reused in an upcoming commit.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>