add test for converting binary relation involving unknown local variables
An upcoming commit will allow nested integer divisions
inside pw_multi_aff_from_map_div.  This means (some of)
the local variables of the binary relation from which
the affine expression is extracted need to be preserved.
However, local variables without a known integer division expression
should still be removed since an affine expression cannot
refer to any such local variables.
Add a test to ensure that this will be handled properly.
The test needs to first parse a binary relation and
then convert it to an affine expression.
The test only needs to check that this conversion succeeds
but actually checks whether the result is the same
as parsing the same string as an affine expression.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>