isl_stream_eat: do not push back unexpected token on stream
The function isl_stream_eat is meant to remove some expected token
from the stream.  If the stream has a different token,
it reports an error and fails.
It also pushes back the unexpected token back onto the stream,
but it should not since this will only result in the same token
getting reported again in isl_stream_free.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>