isl_ast_expr_from_constraint split up affine expression according to sign
That is, instead of constructing expressions based on the sign
of the coefficients, first extract the coefficients of a given sign and
then construct expressions for all coefficients.
This makes the handling of coefficients of opposite sign
more uniform with respect to each other and
also with respect to the handling of coefficients in isl_ast_expr_from_aff.

Reviewed-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>