export isl_id_to_id
This is an example of an export of an isl_*_to_*.
Only the minimal interface to isl_id_to_id that is needed
for testing is exported in this commit.
Further methods can be exported later if needed.

Note that as a side-effect isl_id_to_ast_expr is also exported
because isl/id_to_ast_expr.h is included from isl/ast.h.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>