isl_{set,map}_from_multi_pw_aff: take into account explicit domain
When converting a zero-dimensional isl_multi_pw_aff
to an isl_set or isl_map, the explicit domain of the isl_multi_pw_aff
was not taken into consideration, potentially resulting
in a superset of the expected result.  Make it so that
the resulting set or map is intersected with the explicit domain, if any.

Signed-off-by: Basile Clement <basile.clement@inria.fr>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>