isl_reordering: keep track of target length
This means in particular that the isl_reordering knows
whether isl_reordering_extend has been called so that users
do not need to keep track of this separately.
Even when isl_reordering_extend has not been called,
it makes the target dimensionality more readily available.
This allows some code to be simplified in the next few commits.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>