isl_parameter_alignment_reordering: include other dimensions in reordering
Every call to isl_parameter_alignment_reordering
is followed by a call to isl_reordering_extend_space,
so the reordering constructed by isl_parameter_alignment_reordering
might as well include the other dimensions from the start.

The only exception are union objects, but these have
only parameter dimensions, so they are not affected by this change.
The isl_reordering_extend_space function is still needed there
to extend a reordering on the union object
to reorderings of the different domain spaces.

Reviewed-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>