isl_ast.c: simplify print_sub_expr_c interface
In particular, let print_sub_expr_c obtain the operation and
the subexpression from the operation expression
rather than repeating this in each caller.

Note that the value of the "left" parameter could also be determined
by print_sub_expr_c itself, but here it is easier to do this
in the caller.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>