add isl_*_to_*_read_from_str
Being able to read back associative arrays is useful for debugging and
is required for a proper export of isl_*_to_* to the bindings.
In particular, an upcoming commit will export isl_id_to_id,
adding basic tests on this type.

Make the availability of isl_*_to_*_read_from_str optional
so that external users can opt in.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>