extract_key: allow missing keys in table
The table examined by extract_key is assumed to start at entry 0,
but for a table keyed on enum isl_ast_node_type,
which will be needed in the next commit, there is no entry 0
because the first non-error value is 1 rather than 0.

It is not clear why enum isl_ast_node_type starts at 1, but since
these values appear in public headers, it seems prudent
not to change them.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>