improve isl_pw_aff_list_{min,max} implementation
The original implementation would first combine two elements and
then continue to combine the result with an extra element in each step.
For a list of n elements this results in a subdivisions
into a number of pieces that is in the order of 2^n.
The new implementation directly computes the result
taking into account all elements and only requires a subdivision
into n pieces.

Reviewed-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>