add isl_pw_aff_list_{min,max} test cases
An upcoming commit will change the implementation
of isl_pw_aff_list_{min,max} and there was
no test case that calls these functions
with a list of more than two elements.
Add test cases that end up calling these functions
with a list of four elements.

The representation of the binary relations used
in the test cases is the one that is produced
after the change in implementation.
Any representation would do, but the one currently produced
is larger.

Reviewed-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>