extract out shared isl_ast_node_*_set_*
This will be reused in the next two commits.

Note that the isl_ast_node_*_restore_* functions
are now implemented in terms of isl_ast_node_*_set_*
instead of the other way around because
it is the second variant that will be reused.
The two functions are equivalent, but have a slightly different perspective.
In particular, the "set" variant allows a NULL field for those cases
where a new object is being created and the field has not been set yet,
while the "restore" variant allows a NULL field for those cases
where it was set to NULL by isl_ast_node_*_take_*.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>