properly fix isl_pw_*_gist_params
The previous commit fixed isl_pw_*_gist_params by taking into account
that the intersection of the domain with the context needs
to be specialized for the case where the context is a parameter set.
However, it is this intersection that is later used
to take the gist of the corresponding base expression and
this intersection is always in the domain space.
This means the function to take the gist of the element should be
the same, independently of whether the context is a parameter set.
The isl_*_gist_params was getting used before, which just happens
to work because isl_*_gist_params simply calls isl_set_intersect_params
to get a context in the domain space and then calls isl_*_gist,
while isl_set_intersect_params currently does not check whether
the second argument is indeed a parameter set and simply
performs the same operation as isl_set_intersect.

Note that isl_pw_*_gist_last is only called when one of the domains
is identical to the context, which means that in particular
they live in the same space, so that isl_*_gist can be used there
as well.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>