isl_*_drop_unused_params: extract out shared implementation
This reduces some code duplication and allows the shared implementation
to be reused to implement other isl_*_drop_unused_params functions
in the next commits.

Note that the isl_{basic_,}map_drop name is not entirely
consistent with isl_pw_*_drop_dims.
The generic implementation uses isl_*_drop_dims, which is then
remapped to isl_*_drop for isl_basic_map and isl_map.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>