fix isl_union_pw_qpolynomial_sub
Commit isl-0.11.1-39-ga473a9d0f2 (generalize isl_union_pw_qpolynomial_sub
to include isl_union_pw_multi_aff_sub, Thu Mar 14 09:31:01 2013 +0100)
did not only generalize the original isl_union_pw_qpolynomial_sub,
it also changed the implementation in a way that is valid
for the (then) new isl_union_pw_multi_aff_sub, but
not for isl_union_pw_qpolynomial_sub.
In particular, it resulted in isl_union_pw_qpolynomial_sub producing
an expression that is only defined on the shared domain
of the arguments, while an isl_union_pw_qpolynomial is implicitly zero
on parts of the domain that are not explicitly specified.

Use the original implementation for types with a default zero value.
This implementation is also correct for the other types,
but perhaps slightly less efficient, so keep the "new" implementation
for those other types.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>