isl_input.c: accept_affine_factor: allow integer multiple of identifier
That is, allow an integer multiple of an identifier without
an explicit multiplication.
This allows the next commit to reuse accept_affine_factor
to parse expressions that start with an integer constant.
Note that accept_affine_factor does not currently appear
to get called with a leading integer value token,
so this code path is apparently dead in the current code base,
but it will be revived in the next commit.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>