isl_input.c: extract out next_signed_value*
The users of these functions need a signed value and an upcoming commit
will change the parsing of ISL_TOKEN_VALUE values to no longer
contain any sign.
These users therefore need to handle the sign separately.
The extracted functions will make it easier to perform this handling.

Note that perhaps isl_vec_read_polylib should be changed
to also require the elements of the vector to appear on the same line,
so that only a next_signed_value_on_same_line would be required.
However, it is not clear if allowing coefficients to appear
on different lines was intentional, so it seems safer
to keep the current behavior.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>