no longer treat minus sign as part of token value
The minus sign can be either a unary or a binary operator and
the unary operator binds tighter than the binary one.
Usually the difference is not that important because changing
the sign commutes with several other operations.
However, in the case of integer division, there is a difference
between -(n//d) and (-n)//d and x - n//d should be parsed as x - (n//d),
while it was getting parsed as x + (-n)//d.

Furthermore, -x^y was getting parsed as (-x)^y, while it should
get parsed as -(x^y).  This is an issue when y is even.

Treating -x as two tokens, "-" and "x" makes it easier
to tell the difference between a unary and binary "-" and
to handle -x^y.

The parsing is adjusted to take into account that "-n" now consists
of two tokens.  Note that read_factor already handles negation,
so it does not require any changes.
In other places, the value token is used to represent an exponent,
a denominator, a modulus or the number of rows or variables.
In these cases, negative values are not allowed, so no changes
need to be made to parse them.

Note that the modified code path inside accept_affine_factor
does not appear to be reachable in the current code base.
However, it will get used in an upcoming commit, so
it is also changed to take into account the separate minus sign.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>