isl_input.c: rename extract_mpa_from_tuple to isl_multi_pw_aff_from_tuple
This prepares for the next commit, which extracts
a shared isl_multi_*_from_tuple.
In particular, use a name that can be derived from the type and
that reflects that the "tuple" parameter is freed by the function.
Letting the function free "tuple" allows for (slightly) more code
to be shared.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>