isl_union_*_add_part: properly detect empty parts in result
Empty parts should be removed from the result,
but the code that was meant to perform this check
was looking at one of the addends instead of the sum.
This could result in equal objects that are not obviously equal
to each other simply because one of them has an extra empty part.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>