isl_basic_map_remove_duplicate_constraints: extract opposites up front
The function isl_basic_map_remove_duplicate_constraints
not only looks for equal constraints, but also for opposite constraints.
Separate out the detection of such opposite constraints
from the handling of detected pairs.
This allows the information to be reused in an upcoming commit
at the expense of some extra memory needed to store the information.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>