isl_map_simplify.c: check_for_div_constraints: identify lower bound up front
The function check_for_div_constraints takes in a lower and an upper bound.
The detection can be performed on either bound, but the extraction
is performed on the lower bound.
Identify which of the bounds is the lower bound up front and
perform also the detection on this bound.
This prepares for extracting out a function that does
both the detection and the extraction for a given local variable.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>