isl_map_simplify.c: check_for_div_constraints: consider all local variables
While trying to detect expressions for integer divisions
from a pair of constraints, if the pair initially looks useful
to define the expression for one of the local variables,
but does not end up getting used, the pair would simply
be discarded.

However, the same pair of constraints may involve multiple local variables.
If the pair is not used to define one of them, it may still
be possible to use them to define one of the others.
Continue checking the other local variables instead of
simply giving up.

If the pair does end up getting used, then still no attempt
is made to use the same pair for any other local variable
since that would most likely result in circular definitions and
would therefore get discarded anyway.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>