detect integer divisions based on two pairs of constraints
In particular, given two pairs of constraints

	-n f + g + m n e + c1 >= 0
	n f - g - m n e + c2 >= 0
	g + c3 >= 0
	-g + c4 >= 0

such that

	c1 + c2 < n
	c3 + c4 < m n
	c3 + c4 + (c1 - c3) % n < m n

the variable e is equal to

	floor((f - floor((c1 - c3)/n))/m)

Detect such cases and set the integer division expression of e accordingly,
if it is better than any previously set expression.

The detection of such integer division expressions
can result in tighter simple fixed box hulls,
such as in the new test case.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>