isl_set_get_simple_fixed_box_hull: eliminate local variables involving output
A lower bound on an output dimension cannot involve any local variables
that are unknown or that involve the output dimension itself.
The former group of local variables were already being eliminated,
but not the second.  Do so as well.

Note that a lower bound constraint involving any local variable
currently gets ignored anyway in is_suitable_bound, so
there is no strict requirement for this elimination.
However, the elimination does more than simply dropping or ignoring
the constraints that involve the local variables.
In particular, some linear combinations of the constraints may get added
that can result in tighter boxes.
This is similar to the effect of isl-0.22.1-318-g3f9db650ae
(isl_set_get_simple_fixed_box_hull: eliminate unknown local variables,
Fri Apr 24 14:49:23 2020 +0200).

Furthermore, the next commit will relax the constraints that
can be used to construct a lower bound on an output dimension.
This is more easily done if the constraints involving
inappropriate local variables have already been removed.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>