isl_multi_*_gist: handle explicit domain
That is, if the multi-object has an explicit domain
then take the gist of this domain with respect to the context.
This was arguably missing from isl-0.19-36-g3f2f9ab429 (keep track
of domain in 0D isl_multi_pw_aff and isl_multi_union_pw_aff,
Wed Jun 14 12:05:26 2017 +0200).

Support for explicit domains is added to isl_multi_*_gist
by generalizing the support in isl_multi_*_intersect.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>