isl_*_fixed_power: further simplify intermediate results
In particular, computing a fixed power may result in many
map applications and each of them introduces local variables
that may or may not get simplified away and that may therefore
accumulate as a result of repeated applications.
Try a bit harder to simplify them away by looking
for implicit equality constraints.  This may also be relatively costly,
but if any equality constraints are found, it can significantly reduce
the cost of further applications.
In particular, this simplification is quite effective on the new test case,
where it is actually the coalescing that is partly responsible
for making the constraints more complicated, making it harder
for the local variables to get simplified away automatically.

Reported-by: Jiacheng Pan <jiacheng.pan.sh@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>