isl_set_get_simple_fixed_box_hull: take into account strides
For a set of the form

	[N] -> { [i, N - 4i] : -14 + N <= 16i <= 1 + N }

the first dimension is detected to be equal to floor((1 + N)/16,
i.e., with this offset, the box has size one.
However, if "i" has a fixed value, then clearly N - 4i has a fixed value
as well.  This second fixed value was not being detected,
mainly because the corresponding intermediate set involves
a local variable that is defined in terms of the output dimension and
that is therefore eliminated before looking for a lower bound.

It would be possible to specifically detect equality constraints
that define later dimensions in terms of earlier dimension and
then plug in values of those earlier dimensions if they turn out
to have a fixed value.
However, this would not help if the dimensions are interchanged.

Instead, look for strides in the output dimensions and scale down
the test sets before looking for a lower bound.
This can result in slightly less obvious expressions
(e.g., 4 + N + 4*floor((-2 - N)/16) instead of the perhaps
more obvious N - 4 * floor((1 + N)/16)),
but it works more generally, even in cases where the output dimension
does not have a fixed value.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>