isl_tab_basic_map_partial_lexopt: detect modulos and integer divisions
In particular, if one of the output dimensions is equal
to a modulo or integer division in terms of the parameters and
input dimensions, then there is no need to compute the optimal value
of this output dimension within the lexicographic optimum
since it is known to have a single (parametric) value.

Detecting them beforehand and (essentially) moving them to the context
can avoid some splits of the context as illustrated in the updated
test cases.  It can also result in a different expression
for the modulo or integer division since it is derived
using a different mechanism.

After computing an optimum for the modified problem,
the modulo or integer division is plugged into the solution.
This is achieved using a call to isl_*_pullback_multi_aff.
However, for the isl_map type, there is no function
called isl_map_pullback_multi_aff.  It is called
isl_map_preimage_domain_multi_aff instead.

Note that isl_pw_multi_aff_from_map performs the same detection
of modulos and integer divisions and may end up calling
isl_tab_basic_map_partial_lexopt, which will then perform
the same check.
This can result is some duplicate checks.
Note that isl_pw_multi_aff_from_map performs the check
on the shared constraints over all basic maps,
while isl_tab_basic_map_partial_lexopt does so for each basic map
separately, but if there is only one basic map, the check is the same.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>