isl_tab_basic_map_partial_lexopt*: intersect with domain first
There is only one case where the intersection with the domain
would not be performed before and that is when the total dimensionality
of the domain is zero, but in that case the domain is either empty or
it has no constraints and then the intersection is trivial.

Performing the intersection first allows the remainder of the function
to be extracted out in the next commit.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>