always build Python interface if possible
That is, if the Python interface can be built, then build it
as part of "make all" and not just as part of "make check".
The C++ interfaces are already built as part of "make all"
because they are needed to build the C++ test programs,
while the Python interface is only need to run the Python test program.
Building the Python interface as part or "make all"
prepares for storing the interfaces in the git repository.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>