add generated interfaces
It is usually not a good idea to add generated files
to a git repository.  By definition, a generated file
can be generated from the information that is already available
in the repository.  Furthermore, a generated file may depend
on the machine on which it is generated and so it could be different
for different users.

However, the generated interfaces should specifically be the same
on all systems and if for some reason they are not, then having
them in the repository allows this deviation to be detected.
Furthermore, it is not actually that easy to generate the interfaces
since this requires the clang libraries,
while compiling isl itself just needs any decent C compiler.
Having the interfaces available also means that isl_test2 can
be compiled anywhere where a C++11 compiler is available.
This is especially important since recently
many checks have been moved from (the C) isl_test to (the C++) isl_test2,
while new checks are usually added directly to isl_test2.

Furthermore, keeping track of the interfaces makes it easier
to see the effect of changes on those interfaces and
to see the differences between the interfaces at different points
in history.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>