generate interfaces in source tree
This prepares for the next commit where the generated interfaces
will be stored in the git repository.
In the case of the Python interface, this is only
the machine independent part.

Since "make dist" prefers files in the build tree over those
in the source tree and since the build tree may still contain
stale interfaces, explicitly copy the C++ interfaces in the source tree
in dist-hook.  The full Python interface, including
the machine dependent part is still generated in the build tree,
so it does not need to be copied in this way.

The interfaces stored in the source tree are also temporarily
added to .gitignore.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>